Refactor to improve imports
